move all routes into separate directory
routes/articles
