+++ app.js
@@ -91,85 +91,9 @@ app.get('/', function(req, res) {
   });
 });
 
-// Get Single Article
-app.get('/article/:id', function(req, res) {
-  Article.findById(req.params.id, function(err, article){
-    res.render('article', {
-      article: article
-    })
-    return;
-  });
-});
-
-// Add Route
-app.get('/articles/add', function(req, res){
-  res.render('add_article', {
-    title: 'Add article'
-  });
-})
-
-// Load Edit Form
-app.get('/article/edit/:id', function(req, res) {
-  Article.findById(req.params.id, function(err, article){
-    res.render('edit_article', {
-      title: 'Edit Article',
-      article: article
-    })
-    return;
-  });
-});
-
-// Submit POST Route
-app.post('/articles/add', function(req, res){
-  let article = new Article();
-  article.title = req.body.title;
-  article.author = req.body.author;
-  article.body = req.body.body;
-
-  article.save(function(err){
-    if(err){
-      console.log(err);
-      return;
-    } else {
-      res.redirect('/');
-    }
-    console.log(req.body.title, 'added.')
-  });
-  return;
-});
-
-// Update Submit POST Route
-app.post('/articles/edit/:id', function(req, res){
-  let article = {};
-  article.title = req.body.title;
-  article.author = req.body.author;
-  article.body = req.body.body;
-
-  let query = {_id:req.params.id}
-
-  Article.updateOne(query, article, function(err){
-    if(err){
-      console.log(err);
-      return;
-    } else {
-      res.redirect('/');
-    }
-
-    console.log(req.body.title, 'updated.')
-  });
-});
-
-// Delete article route
-app.delete('/article/:id', function(req, res){
-  let query = {_id: req.params.id}
-
-  Article.deleteOne(query, function(err){
-    if(err){
-      console.log(err);
-    }
-    res.send('Success')
-  });
-});
+// Route Files
+const article = require('./routes/articles')
+app.use('/articles', article)
 
 // Start Server
 app.listen(3000, function() {
